Implement a SoundService, refactor Background music
Implement sound effects. soundFxThread takes soundFx and plays a sound file
associated with that soundFx.
Slightly refactor the way background sound is started/stopped. This is now done
through the SoundService class.
Implemented the "SoundService" class. This class can be used to play the background
music or to play sound effects.
Also added a test GUI for the SoundService.
Added a couple of sound files for sound effects.
Added a LICENSE file, the file contains different licences relating to the system.