Refactor the background music class
The background music class is no longer a thread, this seem to have
been redundant since a clip runs in a daemon thread.

Changed the api of the background music class. Music can now be
toggled on and off through one method call.

Calling playMusic twice will no longer run back the audio clip to
the last paused timestamp.

Updated the SoundService and SoundServiceTestGUI classes to work with
the refactored background music class