Implement dynamic player pieces, and minor changes
Impemented dynamic player pieces with the BoardPiece class. Players now get a board
piece based on their rank and color. When they rank up the piece changes but the color
remains the same.

Selling a property, as well as ranking up, and down now gets logged in the game log.